Add customizable ignored patterns (ex. spaces-only selection)
The default value is set to ignore space-only selections. We can set it to `nil`
by default so we keep the current behavior, with the user being able to
customize it easily.

This PR will fixes the issue addressed by #2 when the value is set to ignore
space-only regions. So, it is a more generic approach to fix this issue.